perf(): do not import bn.js as direct dependency of @zilliqa-js modules
This was causing a significant amount of bloat due to multiple copies of
bn.js ending up in the various packages. This exposes it only from
`@zilliqa-js/util`, which is marked `external` during the bundling
process. It was also causing some confusion, as app developers needed to
independently pull in bn.js on their own.